pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1", # 是一个免费、开源、无需 API 密钥的网络搜索工具
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.30",
    "langchain-milvus>=0.2.0",
    "langchain-openai>=1.1.13",
    "langchain-postgres>=0.0.17",
    "langchain_text_splitters>=1.1.1",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "loguru>=0.7.3",
    "numpy>=2.4.4", # 最基础、最重要的数值计算库
    "nvidia-ml-py>=13.595.45", # 用于获取GPU信息
    "openai>=2.32.0", # 用于创建客户端,调用 API
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pgvector>=0.3.6",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pymilvus==2.6.10",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26", # 它的核心作用是处理 HTTP 请求里的 multipart/form-data 格式数据，最常见的场景就是文件上传。
    "redis>=7.4.0",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1", # 用于加载 Hugging Face 模型(本地模型)
    "spacy>=3.8.14",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0", # 用于创建 Web 应用
    "tenacity>=9.1.4", # 用于重试机制
    "tiktoken>=0.12.0", # 这个 tiktoken 是本地离线计算工具，不用联网、不用 API、不花钱、不泄露你的文本,直接在你电脑上算出准确 Token 数。
    "uvicorn>=0.44.0",
]
